python/2_microservice_sample_app/admin/products/views.py
from rest_framework import viewsets, status
from .models import Product, User
from .serializers import ProductSerializer
from .producer import publish
from rest_framework.response import Response
from rest_framework.views import APIView

import random

class ProductViewSet(viewsets.ViewSet):
    def list(self, request):    # /api/products
        products = Product.objects.all()
        serializer = ProductSerializer(products, many=True)
        return Response(serializer.data)
    

    def create(self, request):  # /api/products
        serializer = ProductSerializer(data=request.data)
        serializer.is_valid(raise_exception=True)
        serializer.save()
        publish('product created', serializer.data)
        return Response(serializer.data, status=status.HTTP_201_CREATED)
     
    def retrieve(self, request, pk=None):  # /api/products/<str:id>
        product = Product.objects.get(id=pk)
        serializer = ProductSerializer(product)
        return Response(serializer.data)
    
    def update(self, request, pk=None):   # /api/products/<str:id>
        product = Product.objects.get(id=pk)
        serializer = ProductSerializer(instance=product, data=request.data)
        serializer.is_valid(raise_exception=True)
        serializer.save()
        publish('product updated', serializer.data)
        return Response(serializer.data, status=status.HTTP_202_ACCEPTED)

    def destroy(self, request, pk=None):  # /api/products/<str:id>
        product = Product.objects.get(id=pk)
        product.delete()
        publish('product deleted', pk)
        return Response(status=status.HTTP_204_NO_CONTENT)

class UserAPIView(APIView):
    def get(self, _):
        users = User.objects.all()
        user = random.choice(users)
        return Response({
            'id': user.id
        })